requirements.txt
Django==2.2.10
psycopg2-binary>=2.8
djangorestframework
django-model-utils
drf-yasg2